+++ web/app/themes/rebe/src/assets/scripts/events/app.js
@@ -2,7 +2,7 @@ angular.module('eventsApp', [
     'ui.router',
     'ngCookies',
     'constants',
-    '720kb.socialshare',  
+    '720kb.socialshare',
     'api.service',
     'event.service',
     'events.timeline.controller',